feat: implementando tratamento de erro customizado para produtos não encontrados
Cria a classe RecursoNaoEncontradoException. Atualiza o ProdutoService para usar orElseThrow e existsById validando os dados. Adiciona bloco try-catch no ProdutoController para capturar a exceção e retornar o status HTTP 404 (Not Found).

+++ src/main/java/com/example/relembrando_conceitos/service/ProdutoService.java
@@ -5,6 +5,7 @@ import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
+import com.example.relembrando_conceitos.exceptions.RecursoNaoEncontradoException;
 import com.example.relembrando_conceitos.model.Produto;
 import com.example.relembrando_conceitos.repository.ProdutoRepository;
 
@@ -21,8 +22,9 @@ public class ProdutoService {
         return produtoRepository.findAll();
     }
 
-    public Optional<Produto> buscaPorId(Long id){
-        return produtoRepository.findById(id);
+    public Produto buscaPorId(Long id){
+        return produtoRepository.findById(id)
+                .orElseThrow(() -> new RecursoNaoEncontradoException("Produto com ID "+id+" não encontrado"));
     }
 
     public Produto salvaProduto(Produto produto){
@@ -30,6 +32,11 @@ public class ProdutoService {
     }
 
     public void deletarProduto(Long id){
+
+        if (!produtoRepository.existsById(id)) {
+            throw new RecursoNaoEncontradoException("Produto com ID "+id+" não encontrado");
+        }
+
         produtoRepository.deleteById(id);
     }
